feat(api/cli): create user cmd
~ refactor click cmds into `commands.py`
